+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 * working Dockerfile
 * capability to stage data for full-slc OPERA RTC run
+* capability to stage data for a burst OPERA RTC run
 * capability to run OPERA RTC from within docker container